refactor: move TestFork into its own module
Relocate the TestFork enum and its chain_spec method from mod.rs into a
dedicated fork.rs, re-exported so test_utils::TestFork is unchanged. This
leaves mod.rs as pure module wiring (declarations + re-exports), with each
concern in its own file: fork, steps, raw_ssz, loader.
